+++ CHANGELOG.md
@@ -8,6 +8,7 @@ ChangeLog
 - IMPROVEMENT: render the "ase-task-list" output as a Markdown table with mtime information
 - REFACTORING: split "task list" functionality into own "ase-task-list" skill
 - CLEANUP: multiple cleanups to various skills
+- UPDATE: upgrade NPM dependencies
 
 0.0.26 (2026-05-13)
 -------------------